+++ lib/build.py
@@ -42,7 +42,7 @@ argfrom = {}
 ## Imports not in standard library ##
 
 # See charliecloud.py for the messy import of this.
-lark = ch.lark
+lark = im.lark
 
 
 ## Main ##
@@ -141,7 +141,7 @@ def main(cli_):
 
    # Count the number of stages (i.e., FROM instructions)
    global image_ct
-   image_ct = sum(1 for i in ch.tree_children(tree, "from_"))
+   image_ct = sum(1 for i in im.tree_children(tree, "from_"))
 
    # Traverse the tree and do what it says.
    #
@@ -256,19 +256,19 @@ class Instruction(abc.ABC):
       self.lineno = tree.meta.line
       self.options = dict()
       # saving options with only 1 saved value
-      for st in ch.tree_children(tree, "option"):
-         k = ch.tree_terminal(st, "OPTION_KEY")
-         v = ch.tree_terminal(st, "OPTION_VALUE")
+      for st in im.tree_children(tree, "option"):
+         k = im.tree_terminal(st, "OPTION_KEY")
+         v = im.tree_terminal(st, "OPTION_VALUE")
          if (k in self.options):
             ch.FATAL("%3d %s: repeated option --%s"
                      % (self.lineno, self.str_name, k))
          self.options[k] = v
 
       # saving keypair options in a dictionary
-      for st in ch.tree_children(tree, "option_keypair"):
-         k = ch.tree_terminal(st, "OPTION_KEY")
-         s = ch.tree_terminal(st, "OPTION_VAR")
-         v = ch.tree_terminal(st, "OPTION_VALUE")
+      for st in im.tree_children(tree, "option_keypair"):
+         k = im.tree_terminal(st, "OPTION_KEY")
+         s = im.tree_terminal(st, "OPTION_VAR")
+         v = im.tree_terminal(st, "OPTION_VALUE")
          # assuming all key pair options allow multiple options
          self.options.setdefault(k, {}).update({s: v})
 
@@ -517,7 +517,7 @@ class Arg(Instruction):
    def __init__(self, *args):
       super().__init__(*args)
       self.commit_files.add(pa.Path("ch/metadata.json"))
-      self.key = ch.tree_terminal(self.tree, "WORD", 0)
+      self.key = im.tree_terminal(self.tree, "WORD", 0)
       if (self.key in cli.build_arg):
          self.value = cli.build_arg[self.key]
          del cli.build_arg[self.key]
@@ -560,9 +560,9 @@ class I_arg_equals(Arg):
    __slots__ = ()
 
    def value_default(self):
-      v = ch.tree_terminal(self.tree, "WORD", 1)
+      v = im.tree_terminal(self.tree, "WORD", 1)
       if (v is None):
-         v = unescape(ch.tree_terminal(self.tree, "STRING_QUOTED"))
+         v = unescape(im.tree_terminal(self.tree, "STRING_QUOTED"))
       return v
 
 
@@ -573,7 +573,7 @@ class Arg_First(Instruction_No_Image):
 
    def __init__(self, *args):
       super().__init__(*args)
-      self.key = ch.tree_terminal(self.tree, "WORD", 0)
+      self.key = im.tree_terminal(self.tree, "WORD", 0)
       if (self.key in cli.build_arg):
          self.value = cli.build_arg[self.key]
          del cli.build_arg[self.key]
@@ -608,9 +608,9 @@ class I_arg_first_equals(Arg_First):
    __slots__ = ()
 
    def value_default(self):
-      v = ch.tree_terminal(self.tree, "WORD", 1)
+      v = im.tree_terminal(self.tree, "WORD", 1)
       if (v is None):
-         v = unescape(ch.tree_terminal(self.tree, "STRING_QUOTED"))
+         v = unescape(im.tree_terminal(self.tree, "STRING_QUOTED"))
       return v
 
 
@@ -641,10 +641,10 @@ class I_copy(Instruction):
          except ValueError:
             pass
       # No subclasses, so check what parse tree matched.
-      if (ch.tree_child(self.tree, "copy_shell") is not None):
-         args = list(ch.tree_child_terminals(self.tree, "copy_shell", "WORD"))
-      elif (ch.tree_child(self.tree, "copy_list") is not None):
-         args = list(ch.tree_child_terminals(self.tree, "copy_list",
+      if (im.tree_child(self.tree, "copy_shell") is not None):
+         args = list(im.tree_child_terminals(self.tree, "copy_shell", "WORD"))
+      elif (im.tree_child(self.tree, "copy_list") is not None):
+         args = list(im.tree_child_terminals(self.tree, "copy_list",
                                              "STRING_QUOTED"))
          for i in range(len(args)):
             args[i] = args[i][1:-1]  # strip quotes
@@ -894,7 +894,7 @@ class I_directive(Instruction_Supported_Never):
 
    @property
    def str_name(self):
-      return "#%s" % ch.tree_terminal(self.tree, "DIRECTIVE_NAME")
+      return "#%s" % im.tree_terminal(self.tree, "DIRECTIVE_NAME")
 
    def prepare(self, *args):
       ch.WARNING("not supported, ignored: parser directives")
@@ -933,10 +933,10 @@ class I_env_equals(Env):
 
    def __init__(self, *args):
       super().__init__(*args)
-      self.key = ch.tree_terminal(self.tree, "WORD", 0)
-      self.value = ch.tree_terminal(self.tree, "WORD", 1)
+      self.key = im.tree_terminal(self.tree, "WORD", 0)
+      self.value = im.tree_terminal(self.tree, "WORD", 1)
       if (self.value is None):
-         self.value = ch.tree_terminal(self.tree, "STRING_QUOTED")
+         self.value = im.tree_terminal(self.tree, "STRING_QUOTED")
 
 
 class I_env_space(Env):
@@ -945,8 +945,8 @@ class I_env_space(Env):
 
    def __init__(self, *args):
       super().__init__(*args)
-      self.key = ch.tree_terminal(self.tree, "WORD")
-      self.value = ch.tree_terminals_cat(self.tree, "LINE_CHUNK")
+      self.key = im.tree_terminal(self.tree, "WORD")
+      self.value = im.tree_terminals_cat(self.tree, "LINE_CHUNK")
 
 
 class I_from_(Instruction):
@@ -982,10 +982,10 @@ class I_from_(Instruction):
       # parent is the last instruction of the base image.
       #
       image_ref = im.Image_Ref(
-         ch.tree_child_terminals_cat(self.tree, "image_ref", "IMAGE_REF"),
+         im.tree_child_terminals_cat(self.tree, "image_ref", "IMAGE_REF"),
          argfrom)
       self.base_image = ch.Image(image_ref)
-      self.alias = ch.tree_child_terminal(self.tree, "from_alias",
+      self.alias = im.tree_child_terminal(self.tree, "from_alias",
                                           "IR_PATH_COMPONENT")
       # Validate instruction.
       if (self.options.pop("platform", False)):
@@ -1096,7 +1096,7 @@ class I_run_exec(Run):
 
    def prepare(self, *args):
       self.cmd = [    ch.variables_sub(unescape(i), self.env_build)
-                  for i in ch.tree_terminals(self.tree, "STRING_QUOTED")]
+                  for i in im.tree_terminals(self.tree, "STRING_QUOTED")]
       return super().prepare(*args)
 
 
@@ -1113,7 +1113,7 @@ class I_run_shell(Run):
       return self._str_  # can't replace abstract property with attribute
 
    def prepare(self, *args):
-      cmd = ch.tree_terminals_cat(self.tree, "LINE_CHUNK")
+      cmd = im.tree_terminals_cat(self.tree, "LINE_CHUNK")
       self.cmd = self.shell + [cmd]
       self._str_ = cmd
       return super().prepare(*args)
@@ -1131,7 +1131,7 @@ class I_shell(Instruction):
 
    def prepare(self, *args):
       self.shell = [    ch.variables_sub(unescape(i), self.env_build)
-                    for i in ch.tree_terminals(self.tree, "STRING_QUOTED")]
+                    for i in im.tree_terminals(self.tree, "STRING_QUOTED")]
       return super().prepare(*args)
 
 
@@ -1148,7 +1148,7 @@ class I_workdir(Instruction):
 
    def prepare(self, *args):
       self.path = pa.Path(ch.variables_sub(
-         ch.tree_terminals_cat(self.tree, "LINE_CHUNK"), self.env_build))
+         im.tree_terminals_cat(self.tree, "LINE_CHUNK"), self.env_build))
       self.chdir(self.path)
       return super().prepare(*args)
 
@@ -1159,7 +1159,7 @@ class I_uns_forever(Instruction_Supported_Never):
 
    def __init__(self, *args):
       super().__init__(*args)
-      self.name = ch.tree_terminal(self.tree, "UNS_FOREVER")
+      self.name = im.tree_terminal(self.tree, "UNS_FOREVER")
 
    @property
    def str_name(self):
@@ -1173,7 +1173,7 @@ class I_uns_yet(Instruction_Unsupported):
 
    def __init__(self, *args):
       super().__init__(*args)
-      self.name = ch.tree_terminal(self.tree, "UNS_YET")
+      self.name = im.tree_terminal(self.tree, "UNS_YET")
       self.issue_no = { "ADD":         782,
                         "CMD":         780,
                         "ENTRYPOINT":  780,
